--- a/.gitignore
+++ b/.gitignore
@@ -1,5 +1,6 @@
 tests/*/results.json
 .DS_Store
+**/.modulecache
 **/.build
 **/Packages
 **/Package.resolved