--- a/Dockerfile
+++ b/Dockerfile
@@ -1,23 +1,29 @@
+# Stage 1: Precompile TestRunner and Module Cache
 FROM swift:6.1.2 AS builder
-# WORKDIR /opt/testrunner
-
-COPY src/TestRunner ./
-
-# Print Installed Swift Version
 RUN swift --version
-#RUN swift package clean
+
+# Build TestRunner executable
+WORKDIR /TestRunner
+COPY src/TestRunner .
 RUN swift build --configuration release
 
+# Build ModuleCache from WarpUm package
+WORKDIR /WarmUp
+COPY src/WarmUp .
+RUN swift build \
+    --build-tests \
+    -Xswiftc -module-cache-path -Xswiftc /opt/test-runner/.modulecache
+
+# Stage 2: Prepare docker container image
 FROM swift:6.1.2
 RUN apt-get update && apt-get install -y jq
+
 WORKDIR /opt/test-runner/
-COPY bin/ bin/
-
-COPY Package.swift ./Package.swift
-
-RUN swift build
-
-COPY --from=builder /.build/release/TestRunner bin/
+COPY bin/run.sh bin/run.sh
+COPY --from=builder /TestRunner/.build/release/TestRunner bin/
+COPY --from=builder /WarmUp/.build .build
+COPY --from=builder /WarmUp/Package.resolved Package.resolved
+COPY --from=builder /opt/test-runner/.modulecache .modulecache/
 
 ENV NAME RUNALL