--- a/Dockerfile
+++ b/Dockerfile
@@ -10,7 +10,7 @@ RUN swift build --configuration release
 # Build WarmUp package
 # Build directory and final working paths should be equal for reuse of ModuleCache.
 # /solution is mounting point where exercise package is added.
-WORKDIR /solution
+WORKDIR /opt/test-runner
 COPY src/WarmUp .
 RUN swift build --build-tests
 
@@ -21,8 +21,8 @@ RUN apt-get update && apt-get install -y jq
 WORKDIR /opt/test-runner/
 COPY bin/run.sh bin/run.sh
 COPY --from=builder /TestRunner/.build/release/TestRunner bin/
-COPY --from=builder /solution/.build .build
-COPY --from=builder /solution/Package.resolved Package.resolved
+COPY --from=builder /opt/test-runner/.build .build
+COPY --from=builder /opt/test-runner/Package.resolved Package.resolved
 
 ENV RUNALL=