--- a/src/WarmUp/Package.swift
+++ b/src/WarmUp/Package.swift
@@ -4,14 +4,26 @@ import PackageDescription
 
 let package = Package(
     name: "WarmUp",
-    products: [],
+    products: [
+        .library(
+            name: "WarmUp",
+            targets: ["WarmUp"]
+        ),
+    ],
     dependencies: [
         .package(url: "https://github.com/apple/swift-numerics", from: "1.0.3")
     ],
     targets: [
+        .target(
+            name: "WarmUp",
+            dependencies: [
+                .product(name: "Numerics", package: "swift-numerics")
+            ]
+        ),
         .testTarget(
             name: "WarmUpTests",
             dependencies: [
+                "WarmUp",
                 .product(name: "Numerics", package: "swift-numerics")
             ]
         )