--- a/src/WarmUp/Tests/WarmUpTests/WarmUpTests.swift
+++ b/src/WarmUp/Tests/WarmUpTests/WarmUpTests.swift
@@ -1,0 +1,6 @@
+import Testing
+import Foundation
+import Numerics
+
+@Test("Stub Test")
+func testStub() { /* Do nothing */ }