Saw the forum [discussion](https://forum.exercism.org/t/anyone-facing-timeout-issues-on-swift-track-exercises/18137/8) about slow build time. Spent some time to research this as no one is planning to do this for now.

There are two major changes that showed significant improvement compared to other approaches I’ve tried:
1. Improved caching logic by warming up all the required dependencies. In the current version of the test runner, this is only done for the Numerics package.
2. Changed the build directory from a Docker-mounted volume to a working directory inside the container. For some reason at least for me it significantly improve build time.

The testing was performed on my local machine using `bin/benchmark-in-docker.sh` and an improved version of `single-test-that-passes`. That changes are not in this PR but, locally I've added various imports such as Foundation and Numerics. The results are as follows:
- Current test runner version: **14.277 s ± 0.549 s**
- Improved cache: **8.787 s ± 0.726 s**
- Improved cache + changed build directory: **6.045 s ± 0.341 s**

## Improving cache

Now I want to briefly guide you through my findings.. To understand what’s happening during the build, you can pass the verbose flag to any Swift command, for example: `swift build -vv`. There’s a lot going on under the hood, with the majority of the work done by swiftc and swift-frontend calls. It turns out that instead of building just two modules—the target under test and the test target—it actually builds four:
- Package
- PackageTests
- PackageDiscoveredTests
- PackagePackageTests (whatever that was)

It looks like PackageDiscoveredTests is related to test auto-discovery. So, maybe that was not the best desicion to get rid of manual test description. But anyway...

This is what happens during the swift-frontend invocation:
```bash
swift test -Xswiftc -stats-output-dir -Xswiftc ./packet.stats -Xswiftc -trace-stats-events -Xswiftc -driver-time-compilation

---User Time---  --System Time--  --User+System--  ---Wall Time--- --- Name ---
0.0100 ( 2.2%)  0.0059 ( 9.5%)  0.0159 ( 3.1%)  1.3145 ( 41.9%) parse-and-resolve-imports
0.0099 ( 2.2%)  0.0059 ( 9.5%)  0.0157 ( 3.1%)  1.3143 ( 41.9%) Import resolution
0.1387 ( 30.8%) 0.0065 ( 10.5%) 0.1452 ( 28.4%) 0.1450 ( 4.6%) build-rewrite-system
0.0448 ( 10.0%) 0.0103 ( 16.5%) 0.0551 ( 10.8%) 0.0551 ( 1.8%) perform-sema
0.0447 ( 9.9%)  0.0103 ( 16.5%) 0.0550 ( 10.7%) 0.0550 ( 1.8%) Type checking and Semantic analysis
0.0446 ( 9.9%)  0.0041 ( 6.7%)  0.0487 ( 9.5%)  0.0487 ( 1.6%) typecheck-stmt
0.0378 ( 8.4%)  0.0044 ( 7.0%)  0.0422 ( 8.2%)  0.0422 ( 1.3%) typecheck-expr
0.0249 ( 5.5%)  0.0099 ( 15.9%) 0.0348 ( 6.8%)  0.0348 ( 1.1%) typecheck-decl
0.0000 ( 0.0%)  0.0040 ( 6.4%)  0.0040 ( 0.8%)  0.0310 ( 1.0%) load-stdlib
0.0285 ( 6.3%)  0.0009 ( 1.4%)  0.0294 ( 5.7%)  0.0295 ( 0.9%) import-clang-decl
0.0172 ( 3.8%)  0.0000 ( 0.0%)  0.0172 ( 3.4%)  0.0172 ( 0.5%) SILGen
0.0147 ( 3.3%)  0.0000 ( 0.0%)  0.0147 ( 2.9%)  0.0147 ( 0.5%) SILGen-function
0.0138 ( 3.1%)  0.0001 ( 0.1%)  0.0139 ( 2.7%)  0.0139 ( 0.4%) precheck-target
0.0086 ( 1.9%)  0.0000 ( 0.0%)  0.0086 ( 1.7%)  0.0086 ( 0.3%) SIL optimization
0.0065 ( 1.4%)  0.0000 ( 0.0%)  0.0065 ( 1.3%)  0.0065 ( 0.2%) IRGen
0.0049 ( 1.1%)  0.0000 ( 0.0%)  0.0049 ( 1.0%)  0.0049 ( 0.2%) typecheck-for-each
0.0001 ( 0.0%)  0.0000 ( 0.0%)  0.0001 ( 0.0%)  0.0001 ( 0.0%) source-file-populate-cache
0.0001 ( 0.0%)  0.0000 ( 0.0%)  0.0001 ( 0.0%)  0.0001 ( 0.0%) get-conformance-access-path
0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%) SIL verification, pre-optimization
0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%) populate-source-file-class-member-cache
0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%) SIL verification, post-optimization
0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%) module-populate-cache
0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%) AST verification
0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%)  0.0000 ( 0.0%) perform-whole-module-type-checking
0.4498 (100.0%) 0.0622 (100.0%) 0.5120 (100.0%) 3.1361 (100.0%) Total
```
Almost 3 seconds is wasted on parsing, resolving imports, and import resolution. It turns out that during this time, swift-frontend builds all the internal dependencies (clang and swift modules) required for the files for each package. Then can be seen in ModuleCache directory.

So the idea of this one is two create a package called WarmUp, place all the imports. Then compile and store .build directory into docker image. It's already done for Numerics package. Everything else is added in this PR.

## Changing build directory
It turns out building inside docker image is more efficient then working with mounted directory.