@Sencudra 2025-07-30
@meatball133 Take as much time as you need.

A few words about the restructuring. Before this PR, the folder with the exercise was supposed to be mounted into the Docker image. That part hasn’t changed — except now, instead of building in the mounted folder, it copies all the contents (the exercise package) into the Docker image and builds it there, inside. Tests showed an improvement — at least on my machine.

So, here is the question. The old run-test-docker.sh ran all the tests in the same Docker image, so to work it now it's necessary to have a script that removes artifacts between test runs. Not only to make it work properly, but also to ensure that the benchmark is accurate and independent between runs. So, between deleting artifacts manually and running a clean Docker image, I chose the latter, since deletion didn't look trivial. You're supposed to delete all the artifacts while preserving things like Package.resolved, bin/, etc. Not knowing that all the runners follow the same file structure, I went with the approach to just use a clean docker for each test. That's why scripts have changed.

> I don't really see the need for such a heavy restructure, even if yes it would save on execution time.

That would be a valid point — if the runner didn’t hit Docker timeouts. Overall, I believe the need to preserve the file structure pattern in the repository should not outweigh the effectiveness and simplicity of the code.